requirements.txt
alembic==1.4.2
Babel==2.8.0
blinker==1.4
click==7.1.2
Flask==1.1.2
Flask-Migrate==2.5.3
Flask-Moment==0.9.0
Flask-SQLAlchemy==2.4.1
Flask-WhooshAlchemy==0.56
Flask-WTF==0.14.3
itsdangerous==1.1.0
Jinja2==2.11.2
Mako==1.1.2
MarkupSafe==1.1.1
psycopg2==2.8.5
python-dateutil==2.6.0
python-editor==1.0.4
pytz==2020.1
six==1.14.0
SQLAlchemy==1.3.16
Werkzeug==1.0.1
Whoosh==2.7.4
WTForms==2.3.1
